+++ crates/zng-app/src/running.rs
@@ -7,14 +7,14 @@ use std::{
     time::{Duration, Instant},
 };
 
-use crate::{Deadline, view_process::raw_device_events::INPUT_DEVICES};
+use crate::Deadline;
 use parking_lot::Mutex;
 use zng_app_context::{AppScope, app_local};
 use zng_task::DEADLINE_APP;
 use zng_time::{INSTANT_APP, InstantMode};
 use zng_txt::Txt;
-use zng_var::{ArcVar, ReadOnlyArcVar, ResponderVar, ResponseVar, VARS, VARS_APP, Var as _, response_var};
-use zng_view_api::raw_input::InputDeviceEvent;
+use zng_var::{ArcVar, ReadOnlyArcVar, ResponderVar, ResponseVar, VARS_APP, Var as _, response_var};
+use zng_view_api::{DeviceEventsFilter, raw_input::InputDeviceEvent};
 
 use crate::{
     APP, AppControlFlow, AppEventObserver, AppExtension, AppExtensionsInfo, DInstant, INSTANT,
@@ -77,11 +77,10 @@ impl<E: AppExtension> RunningApp<E> {
             let _t = INSTANT_APP.pause_for_update();
             extensions.register(&mut info);
         }
-        let device_events = extensions.enable_input_device_events();
 
         {
             let mut sv = APP_PROCESS_SV.write();
-            sv.set_extensions(info, device_events);
+            sv.set_extensions(info);
         }
 
         if with_renderer && view_process_exe.is_none() {
@@ -93,7 +92,7 @@ impl<E: AppExtension> RunningApp<E> {
         #[cfg(target_arch = "wasm32")]
         let view_process_exe = std::path::PathBuf::from("<wasm>");
 
-        let process = AppIntrinsic::pre_init(is_headed, with_renderer, view_process_exe, view_process_env, device_events);
+        let process = AppIntrinsic::pre_init(is_headed, with_renderer, view_process_exe, view_process_env);
 
         {
             let _s = tracing::debug_span!("extensions.init").entered();
@@ -585,16 +584,6 @@ impl<E: AppExtension> RunningApp<E> {
                 zng_view_api::Event::Inited(zng_view_api::Inited {
                     generation,
                     is_respawn,
-                    available_monitors,
-                    available_input_devices,
-                    multi_click_config,
-                    key_repeat_config,
-                    touch_config,
-                    font_aa,
-                    animations_config,
-                    locale_config,
-                    colors_config,
-                    chrome_config,
                     extensions,
                     ..
                 }) => {
@@ -606,34 +595,8 @@ impl<E: AppExtension> RunningApp<E> {
 
                     VIEW_PROCESS.handle_inited(generation, extensions.clone());
 
-                    let monitors: Vec<_> = available_monitors
-                        .into_iter()
-                        .map(|(id, info)| (VIEW_PROCESS.monitor_id(id), info))
-                        .collect();
-
-                    VARS.animations_enabled().set(animations_config.enabled);
-
-                    let args = crate::view_process::ViewProcessInitedArgs::now(
-                        generation,
-                        is_respawn,
-                        monitors,
-                        multi_click_config,
-                        key_repeat_config,
-                        touch_config,
-                        font_aa,
-                        animations_config,
-                        locale_config,
-                        colors_config,
-                        chrome_config,
-                        extensions,
-                    );
+                    let args = crate::view_process::ViewProcessInitedArgs::now(generation, is_respawn, extensions);
                     self.notify_event(VIEW_PROCESS_INITED_EVENT.new_update(args), observer);
-
-                    let devices: HashMap<_, _> = available_input_devices
-                        .into_iter()
-                        .map(|(d_id, info)| (self.input_device_id(d_id), info))
-                        .collect();
-                    INPUT_DEVICES.update(devices);
                 }
                 zng_view_api::Event::Suspended => {
                     VIEW_PROCESS.handle_suspended();
@@ -1239,13 +1202,7 @@ struct PendingExit {
 }
 impl AppIntrinsic {
     /// Pre-init intrinsic services and commands, must be called before extensions init.
-    pub(super) fn pre_init(
-        is_headed: bool,
-        with_renderer: bool,
-        view_process_exe: PathBuf,
-        view_process_env: HashMap<Txt, Txt>,
-        device_events: bool,
-    ) -> Self {
+    pub(super) fn pre_init(is_headed: bool, with_renderer: bool, view_process_exe: PathBuf, view_process_env: HashMap<Txt, Txt>) -> Self {
         APP_PROCESS_SV
             .read()
             .pause_time_for_updates
@@ -1265,12 +1222,12 @@ impl AppIntrinsic {
             debug_assert!(with_renderer);
 
             let view_evs_sender = UPDATES.sender();
-            VIEW_PROCESS.start(view_process_exe, view_process_env, device_events, false, move |ev| {
+            VIEW_PROCESS.start(view_process_exe, view_process_env, false, move |ev| {
                 let _ = view_evs_sender.send_view_event(ev);
             });
         } else if with_renderer {
             let view_evs_sender = UPDATES.sender();
-            VIEW_PROCESS.start(view_process_exe, view_process_env, false, true, move |ev| {
+            VIEW_PROCESS.start(view_process_exe, view_process_env, true, move |ev| {
                 let _ = view_evs_sender.send_view_event(ev);
             });
         }
@@ -1297,7 +1254,14 @@ impl AppIntrinsic {
 }
 impl AppExtension for AppIntrinsic {
     fn event_preview(&mut self, update: &mut EventUpdate) {
-        if let Some(args) = EXIT_CMD.on(update) {
+        if VIEW_PROCESS_INITED_EVENT.has(update) {
+            let filter = APP_PROCESS_SV.read().device_events_filter.get();
+            if !filter.is_empty() {
+                if let Err(e) = VIEW_PROCESS.set_device_events_filter(filter) {
+                    tracing::error!("cannot set device events on the view-process, {e}");
+                }
+            }
+        } else if let Some(args) = EXIT_CMD.on(update) {
             args.handle_enabled(&self.exit_handle, |_| {
                 APP.exit();
             });
@@ -1305,7 +1269,13 @@ impl AppExtension for AppIntrinsic {
     }
 
     fn update(&mut self) {
-        if let Some(response) = APP_PROCESS_SV.write().take_requests() {
+        let mut sv = APP_PROCESS_SV.write();
+        if let Some(filter) = sv.device_events_filter.get_new() {
+            if let Err(e) = VIEW_PROCESS.set_device_events_filter(filter) {
+                tracing::error!("cannot set device events on the view-process, {e}");
+            }
+        }
+        if let Some(response) = sv.take_requests() {
             let args = ExitRequestedArgs::now();
             self.pending_exit = Some(PendingExit {
                 handle: args.propagation().clone(),
@@ -1379,7 +1349,7 @@ app_local! {
     pub(super) static APP_PROCESS_SV: AppProcessService = AppProcessService {
         exit_requests: None,
         extensions: None,
-        device_events: false,
+        device_events_filter: zng_var::var(Default::default()),
         pause_time_for_updates: zng_var::var(true),
         is_suspended: zng_var::var(false),
     };
@@ -1388,7 +1358,7 @@ app_local! {
 pub(super) struct AppProcessService {
     exit_requests: Option<ResponderVar<ExitCancelled>>,
     extensions: Option<Arc<AppExtensionsInfo>>,
-    pub(super) device_events: bool,
+    pub(crate) device_events_filter: ArcVar<DeviceEventsFilter>,
     pause_time_for_updates: ArcVar<bool>,
     is_suspended: ArcVar<bool>,
 }
@@ -1414,9 +1384,8 @@ impl AppProcessService {
             .unwrap_or_else(|| Arc::new(AppExtensionsInfo { infos: vec![] }))
     }
 
-    pub(super) fn set_extensions(&mut self, info: AppExtensionsInfo, device_events: bool) {
+    pub(super) fn set_extensions(&mut self, info: AppExtensionsInfo) {
         self.extensions = Some(Arc::new(info));
-        self.device_events = device_events;
     }
 }