+++ crates/zng-view-api/src/lib.rs
@@ -272,11 +272,16 @@ declare_api! {
     ///
     /// Sends an [`Event::Inited`] once the view is completely connected.
     /// Other methods may only be called after this event.
-    fn init(&mut self, vp_gen: ViewProcessGen, is_respawn: bool, device_events: bool, headless: bool);
-    // TODO(breaking) refactor device_events, add audio_events
+    fn init(&mut self, vp_gen: ViewProcessGen, is_respawn: bool, headless: bool);
 
     /// Called once after exit, if running in a managed external process it will be killed after this call.
     fn exit(&mut self);
+
+    /// Enable/disable global device events.
+    ///
+    /// This filter affects device events not targeted at windows, such as mouse move outside windows or
+    /// key presses when the app has no focused window.
+    pub fn set_device_events_filter(&mut self, filter: DeviceEventsFilter);
 
     /// Open a window.
     ///