+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # Unreleased
 
-* **Breaking** Renamed `AppExtension::enable_device_events` to `enable_input_device_events`.
-    - Also renamed `APP.device_events` `APP.input_device_events`.
+* **Breaking** Refactored how view-process config events notify.
+    - Initial non default config state now reported as events on init.
+    - All config and monitors info removed from `ViewProcessInitedArgs` and related API.
+* Fix  `VARS.animations_enabled` not updating when it is not set and the `sys_animations_enabled` changes. 
+
+* **Breaking** Refactored how raw device events are enabled on the view-process.
+    - Now can dynamically enable/disable and with more precision of what kind of events.
+    - Removed `enable_input_device_events` and all related API from view-process controller.
+    - Added `APP.device_events_filter` variable that can be set to enable/disable device events.
+
 * **Breaking** Remove all deprecated items.
+
 * **Breaking** Refactor `zng::slider` API.
     - Removed direct support to std range type, use `Selector::many` with two values.
     - Selector `value_with` and `many_with` now expects `Sync` closures.